requirements.txt
streamlit
numpy
matplotlib
pillow
opencv-python
imageio
pygame